Base label color on disabled prop
This commit alters the color of labels for both the Radio and Checkbox
components if the disabled prop is sent down as true.

The idea behind this is that the visual cue that indicates the
component is disabled becomes more reliable when it is applied to the
element as a whole.