Add children to Dropdown props interface
While attempting to import the `<Dropdown>` component from Flame in the new insights app (https://github.com/lightspeed-hospitality/lighthouse-insights), we ran into a Typescript issue where the app wouldn't compile because `children` was not a valid prop in the `Dropdown` interface.

The insights app is using React 18 and in that version React.FC no longer includes children by default.
- https://stackoverflow.com/a/71809927/16826114
- https://solverfox.dev/writing/no-implicit-children/

So any Flame components that use React.FC and a children prop will probably need to be updated to be compatible with React 18 and TS. For now we're making this one change in this one component we need as a way to kick the tires on contributing to Flame and to make sure it works as we expect.