Revert: remove children from Dropdown type (#172)
* Revert: remove children from Dropdown type (#171)

The new typing of `React.node` is too strict. The Dropdown can accept other children types, like a
child function that contains the close event as first parameter, as seen in this first example:
https://lightspeed-flame.netlify.app/?path=/story/components-dropdown--story.

Instead of trying to find a better more flexible and specific type, remove it for now and revisit
later. We can look at this in context of upgrading the entire package for React 18 compatibility.

* Update docs re: yarn link and releasing to next

* Typo

* Update .github/CONTRIBUTING.md

Co-authored-by: Guillaume Lambert <24464151+glambert@users.noreply.github.com>

Co-authored-by: Guillaume Lambert <24464151+glambert@users.noreply.github.com>